+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 * 1.0.0-beta8 (@release_date@)
 
  * added `--class` option to `puli bind --update` command
+ * added `puli upgrade` command
  
 * 1.0.0-beta7 (2015-08-24)