+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-* 1.0.0-beta8 (@release_date@)
+* 1.0.0-beta8 (2015-10-05)
 
  * changed `puli bind` command to generate `ClassBinding` instances when passing
    a path that contains backslashes (`\`) or when passing the `--class` option